add HDF test unit
